requirements.txt
beautifulsoup4==4.12.2
certifi==2023.5.7
chardet==3.0.4
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
deep-translator==1.11.1
googletranslate-python==1.0.3
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
markdown-it-py==2.2.0
mdurl==0.1.2
polib==1.2.0
progress==1.6
Pygments==2.15.1
requests==2.31.0
rfc3986==1.5.0
rich==13.3.5
shellingham==1.5.0.post1
sniffio==1.3.0
soupsieve==2.4.1
tqdm==4.65.0
typer==0.9.0
typing_extensions==4.6.2
urllib3==2.0.2
